fastapi_auth_service/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
psycopg2-binary
pydantic
